refactor:change imports after moving HttpError from type.ts to index.ts
