+++ lib/datasource/pod/index.ts
@@ -3,8 +3,8 @@ import { HOST_DISABLED } from '../../constants/error-messages';
 import { logger } from '../../logger';
 import { ExternalHostError } from '../../types/errors/external-host-error';
 import { cache } from '../../util/cache/package/decorator';
+import type { HttpError } from '../../util/http';
 import { GithubHttp } from '../../util/http/github';
-import type { HttpError } from '../../util/http/types';
 import { newlineRegex, regEx } from '../../util/regex';
 import { Datasource } from '../datasource';
 import { massageGithubUrl } from '../metadata';
